+++ nrv/utils/cell/CL_simulations.py
@@ -2,8 +2,10 @@
 NRV-Cellular Level simulations.
 """
 import sys
+from typing import Callable
+from time import perf_counter
 
-from ...backend.log_interface import pass_info, rise_error, rise_warning
+from ...backend.log_interface import pass_info, rise_error, rise_warning, clear_prompt_line
 from ...backend.MCore import *
 from ...fmod.electrodes import *
 from ...fmod.extracellular import *
@@ -25,6 +27,118 @@ unmyelinated_models = [
 ]
 myelinated_models = ["MRG", "Gaines_motor", "Gaines_sensory"]
 
+
+
+def axon_AP_threshold(axon: axon, amp_max: float, update_func: Callable,
+                      t_sim: float = 5, tol: float = 1,
+                      verbose: bool = True, args_update = None ,**kwargs) -> np.float64 :
+    """
+    Find the activation threshold of an axon with arbitrary stimulation settings.
+
+    Parameters
+    ----------
+    axon : axon
+        axon object to simulation
+    amp_max : float
+        maximum amplitude for the binary search, in µA
+    update_func : Callable
+        Callable function to update the axon stimulation parameters between each binary search iteration
+    t_sim : float, optional
+        Simulation duration, in ms, by default 5
+    tol : float, optional
+        Search tolerance, in %, by default 1
+    verbose : bool, optional
+        Verbosity of the search, by default True
+    args_update : arg, optional
+        update_func arguments, by default None
+
+    Returns
+    -------
+    current_amp : np.float64
+        The threshold found with the binary search.
+    """
+
+    if "amp_tol_abs" in kwargs:
+        amp_tol_abs = kwargs.get("amp_tol_abs")
+    else:
+        amp_tol_abs = 0
+
+    if "amp_min" in kwargs:
+        amp_min = kwargs.get("amp_min")
+    else:
+        amp_min = 0
+    
+    amplitude_max_th = amp_max
+    amplitude_min_th = amp_min
+    amplitude_tol = tol
+
+    # Dichotomy initialization
+    previous_amp = amp_min
+    delta_amp = np.abs(amp_max - amp_min)
+    current_amp = amp_max
+    Niter = 1
+    keep_going = 1
+    t_start = perf_counter()
+
+    while keep_going:
+        if verbose and Niter == 1:
+            pass_info(f"Iteration {Niter}, Amp is {np.round(current_amp,2)}µA ...")
+            
+        update_func(axon,current_amp,**args_update)
+        results = axon.simulate(t_sim=t_sim)
+
+        if verbose:
+            clear_prompt_line(1)
+
+        # post-process results
+        delta_amp = np.abs(current_amp - previous_amp)
+        #put in a function
+        if amp_tol_abs > 0:
+            if delta_amp < amp_tol_abs:
+                keep_going = 0
+            else:
+                keep_going = 1
+        else:
+            if current_amp > previous_amp:
+                current_tol = 100 * delta_amp / current_amp
+            else:
+                current_tol = 100 * delta_amp / previous_amp
+            if current_tol <= tol:
+                keep_going = 0
+            else:
+                keep_going = 1
+        previous_amp = current_amp
+        # test simulation results, update dichotomy
+        if results.is_recruited():
+            if current_amp == amp_min:
+                rise_warning("Minimum Stimulation Current is too High!")
+                return 0
+            if verbose:
+                pass_info(f"Iteration {Niter}, Amp is {np.round(current_amp,2)}µA"+ 
+                          f" ({np.round(current_tol,2)}%)... AP Detected! (in {np.round(results["sim_time"],3)}s)")
+                #pass_info("... Spike triggered")
+            amplitude_max_th = previous_amp
+            current_amp = (delta_amp / 2) + amplitude_min_th
+        else:
+            if current_amp == amp_max:
+                rise_warning("Maximum Stimulation Current is too Low!")
+                return 0
+            if verbose:
+                pass_info(f"Iteration {Niter}, Amp is {np.round(current_amp,2)}µA"+
+                          f" ({np.round(current_tol,2)}%)... AP Not Detected! (in {np.round(results["sim_time"],3)}s)")
+            current_amp = amplitude_max_th - delta_amp / 2
+            amplitude_min_th = previous_amp
+
+        if previous_amp == amp_max:
+            current_amp = amp_min
+
+        Niter += 1
+    t_stop = perf_counter()
+    if verbose:
+        clear_prompt_line(1)
+        pass_info(f"Activation threshold is {np.round(current_amp,2)}µA ({np.round(current_tol,2)}%),"
+                  + f" found in {Niter-1} iterations ({np.round(t_stop-t_start,2)}s).")
+    return current_amp
 
 def firing_threshold_point_source(
     diameter,
@@ -104,6 +218,11 @@ def firing_threshold_point_source(
     threshold       : float
             estimated firing threshold in uA
     """
+
+    rise_warning(
+        "DeprecationWarning: ",
+        "firing_threshold_point_source is obsolete use axon_AP_threshold instead"
+    )
 
     if "t_sim" in kwargs:
         t_sim = kwargs.get("t_sim")
@@ -374,6 +493,11 @@ def firing_threshold_from_axon(
     threshold       : float
             estimated firing threshold in uA
     """
+
+    rise_warning(
+        "DeprecationWarning: ",
+        "firing_threshold_point_source is obsolete use firing_threshold_from_axon instead"
+    )
 
     if "elec_id" in kwargs:
         elec_id = kwargs.get("elec_id")